Module Equal2.Extend

Interface builder for equatable polymorphic binary types.

Parameters

Base : Equal2

Signature

type ('a, 'b) t

Abstract type for equatable values.

val equal : 'a equality -> 'b equality -> ('a, 'b) t equality

eq equal_a equal_b t1 t2 tests if the values t1 and t2 are equal using the equal_a and equal_b functions to compare the contained values of type 'a and 'b respectively.

val not_equal : 'a equality -> 'b equality -> ('a, 'b) t equality

not_eq equal_a equal_b t1 t2 tests if the values t1 and t2 are not equal using the equal_a and equal_b functions to compare the values of type 'a and 'b respectively.

not_equal equal_a equal_b t1 t2 is a shorthand for not (equal equal_a equal_b t1 t2).